Optimize deck simulator: parallelism, double math, attribution gating
Performance work on the Monte Carlo deck simulator and play-value tooling:

- Expected-value runs (SimulateExpectedTurnValues) build no card-value
  attribution; PlayValue computes modifier value numerically and emits credit
  events only when attribution is collected. EV and search are unchanged.
- Simulate takes an optional TrackedCreditModelId so direct play-value probes
  retain only the probe card's credit rows.
- Simulate parallelizes its run loop (per-thread run seeds, sequential
  accumulation -> identical results to the serial path).
- Clone skips empty ForgeCredits allocation and pre-sizes pile lists.
- FastRandom (SplitMix64) replaces System.Random on the hot path: deterministic
  but a different stream from the legacy generator.
- Inner simulation math is double; decimal survives only at the report boundary
  (DeckSimulationReport, runtime value JSON, static CardValueEstimator).
  SimulationCard numeric fields and the scorer interface are double.

Tooling/CLI:
- Estimation parallelism defaults to 4 cores (--degree-of-parallelism); new
  --run-degree (default 4) parallelizes a single deck's runs for
  estimate-direct-play-values / estimate-floor8-play-values, guarded so it does
  not nest with the per-card layer. simulate-deck-scenario gains --run-degree.

Docs/skills:
- Document the probe-strategy principle: fully attributable probes use
  source-credit; probes with a non-numerically-attributable term (notably draw,
  e.g. BigBang) use play-delta (normal vs blocked). Add simulation-performance
  notes to AGENTS.md, CLAUDE.md, README.md and the direct-play/eligibility skills.

Core.Tests and Modeling.Tests pass with no value drift.

+++ CardValueOverlay.Modeling/Simulation/NeuralSearchCardScorer.cs
@@ -58,7 +58,7 @@ public sealed class NeuralSearchCardScorer : ISearchCardScorer
         return new NeuralSearchCardScorer(model);
     }
 
-    public decimal Score(SearchCardScoringContext context)
+    public double Score(SearchCardScoringContext context)
     {
         double[] input = BuildInput(context);
         double[] values = input;
@@ -72,7 +72,7 @@ public sealed class NeuralSearchCardScorer : ISearchCardScorer
             throw new InvalidOperationException($"Search policy model output length {values.Length} is invalid; expected 1.");
         }
 
-        return (decimal)values[0];
+        return (double)values[0];
     }
 
     private double[] BuildInput(SearchCardScoringContext context)
